refactor: extract ActiveRecordRelationMethods to new file
